[FIX] fix compile error
- resolve overloaded call to log function
- fix MSVS error

+++ src/utils/SimpleSearchEngine.cpp
@@ -330,7 +330,7 @@ class SimpleSearchEngine :
         double z = 0;
         for (y = 2; y <= x; y++)
         {
-          z = log(y) + z;
+          z = log((double)y) + z;
         }
 
         return z;
